feat: add missing cookie parameter
Newer versions of electron require SameSite=none to be explicitly set in the cookie.
The best way would be to change the server to include this parameter or make sure that requests come
from the same domain.
In the meantime, this is a good enough workaround and allows us to use the latest electron version